refactor(desktop): split monoliths, fix startup performance
Frontend JS (2950 → 2452 lines + 3 modules):
- Extract bridge.js: Tauri invoke/listen wrappers with mock fallback
- Extract utils.js: pure utility functions (delay, nowTs, fmtNumber, fmtTs, escapeHtml, parseGitHubRepoName)
- Extract mock.js: MockState, mockInvoke, browser simulation
- main.js now imports from modules; no circular deps

Rust backend (1765 → 64 lines + 4 modules):
- types.rs: all type definitions (Task, ChatMessage, Config, BridgeCommand, BridgeEvent, etc.)
- state.rs: AppState + persistence (load/save/sync functions)
- bridge_loop.rs: start_bridge, event handling, orchestration runner
- commands.rs: all 29 #[tauri::command] handlers
- main.rs: entry point only (mod declarations + fn main)

Startup performance fixes:
- backend.rs: 10 → 5 retries, 3s → 1s initial delay, 30s → 10s max backoff
- backend.rs: health check interval 10s → 30s, reconnect 5 → 3 retries
- windows.rs: skip WSL kill/restart if bridge already reachable (warm launch)
- windows.rs: wait_for_bridge timeout 90s → 25s (fail fast, retry sooner)
- windows.rs: detect_wsl2 capped at 8s timeout to prevent wsl.exe hangs
